Use custom action to deploy to pages

+++ dev/before_build.sh
@@ -2,5 +2,7 @@
 
 # Execute this script in the root directory of the project!
 
-echo "Removing ./dist/ contents before building"
-rm -r ./dist/*
+if [ -d "./dist" ]; then
+    echo "Removing ./dist/ contents before building"
+    rm -r ./dist/*
+fi
